``Field`` instances are hashable on Python 3
They use a defined hashing algorithm that matches what equality does
on all versions of Python. Previously, on Python 2, fields were hashed
based on their identity. This violated the rule that equal objects
should have equal hashes, and now they do. Since having equal hashes
does not imply that the objects are equal, this is not expected to be
a compatibility problem.

Included tests with a dict show that there shouldn't be false matches
even with equal hashes.

Fixes #36