Allow `missing_value` without setting `min` and `max`
Fixes #9

Note that you must still specify a `default` in that case. I'm not
sure that makes sense or not.

This impacts Int, Float, Decimal, DateTime, Timedelta, Date, and
Time---all the subclasses of Orderable. They all have a test case.

Also fix ValidatedProperty and DefaultProperty to do the right thing
when accessesd on the class and not the instance (which is to return
themself; previously they would raise AttributeError).