Handle Unicode token values with non-ascii chars.
Fixes #15 and fixes #6

I went with backslashreplace because I couldn't find anything that was portable and operated on textual data (binascii and quopri need binary data; namereplace needs Python 3.5+).

The ITerm comment claims that token should be an IASCIILine, which is a native string, not necessarily text, so that's what I kept it here.

Note that we're not actually enforcing that the value is an IASCIILine. Passing in control characters would still be allowed as before.