Add fields and interfaces matching the numeric tower.
In descending order of generality these are ``Number``, ``Complex``,
``Real``, ``Rational`` and ``Integral``. The ``Int`` class extends
``Integral``, the ``Float`` class extends ``Real``, and the
``Decimal`` class extends ``Number``.

Generalize the parsing of numbers in the Number class.

Add tests and document them.

Run the field and bootstrap field doctests as part of unittesting, not
just under tox with sphinx, and make them run on both Python 2 and 3.

Fixes #49