Make Object a bootstrapfield and share the logic between Object validation and the public functions get[Schema]ValidationErrors.
Fixes #57

This makes Object bind all fields (not just choices) before
validation. This fixes #17 in a backwards compatible way.

Binding all attributes, not just Choices, reduced the dependencies of
Object and facilitated making it a bootstrap field. Making it a
bootstrap field in turn let us use it in more places in interfaces.py,
which fixes #13.

Switch from just repeating attribute names in interfaces.py to a real
__all__ attribute that linters can warn about.

Change ..autoclass:: in api.rst to ..autointerface:: so we get the
actual member documentation and not lots of warnings about missing
__mro__. (This is unrelated, I was just tired of the warnings.)