Add IFromBytes
As a counterpoint to IFromUnicode.

This is useful when you know the type of data you're dealing with, and conversions are either trivially non-lossy (or will fail anyway) or are not necessary.

The numeric fields as well as Bytes/Line and URI, DottedName and Id implement this, the last three because they are native string lines that want to accept bytes on Python 2 anyway.

On Python 2, lxml likes to produce byte strings "to save memory" even when parsing unicode and given proper XML encoding information, so this can be very helpful there.