Add human-readable __str__ to ValidationError
When upgraded to zope.schema >= 4.6.0, one of Launchpad's doctests
failed with this output, from an API endpoint that printed validation
errors on a field named 'tag' when attempting to set it to an invalid
value:

    - tag: an invalid tag !!!
    + tag: (u'an invalid tag !!!', 'tag')
    ?      +++                  +++++++++

This is produced by lazr.restful, using code along the lines of:

    try:
        field.validate(value)
    except ValidationError as e:
        errors.append(u"%s: %s" % (name, e))

While I can obviously just change the doctest, neither the old nor the
new output is particularly good here.  The old output is neat, but it
doesn't actually say what the problem is other than giving the field
name and the objectionable value (is it an unsatisfied constraint, or a
wrong type, or some kind of nested validation error, or ...).  The new
output is worse: it exposes too much syntax for a good error message,
and it duplicates the field name because it's just the stringification
of the tuple of all the exception arguments.  Switching to the
exception's `repr` would fix the issue that we don't say what the
problem is, but in a way that's made up of Python syntax and is intended
for developers rather than users.

Although I haven't specifically run into them yet, we may well have
similar problems in our web UI.  In that case it will definitely be
inappropriate to show the field name (since that would be an identifier
that doesn't necessarily correspond to the field's label in the UI), and
again we'll want to avoid excessive Python syntax leaking out.

While we could work around this in lazr.restful and/or Launchpad, it
seems better to fix this at its root and give ValidationError instances
a proper `__str__` implementation.  The principles I've tried to follow
are:

 * Don't render field names, except in some cases of nested validation
   errors; the frontend is likely to have a better idea of how to
   indicate which error goes with which field.

 * State the nature of the validation error as well as just its
   arguments.

 * Minimise code duplication, to make it likely that if somebody adds
   another ValidationError subclass it will be stringified reasonably by
   default.

The results aren't perfect: I'd hoped to try to make the `__str__`
output be a single line in more cases, but the existing docstrings make
that a bit awkward to achieve without parsing translated strings in a
way I don't think would be appropriate.  The rendering of nested errors
might also get a bit hard to read in extreme cases.  Nevertheless, I
think this is a much better basis for frontends displaying errors to
users.